requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==39.0.2
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.8.0
django-crispy-forms==2.0
django-redis==5.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.2.0
kombu==5.2.4
msgpack==1.0.5
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.38
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-decouple==3.7
pytz==2022.7.1
redis==4.3.4
regex==2022.10.31
requests==2.28.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.2
torch==1.13.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==5.5.2
